requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin-material-dashboard==1.0.10
django-auditlog==3.0.0
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
